+++ package.json
@@ -28,10 +28,12 @@
     "monaco-editor": "^0.19.3",
     "node-fetch": "^2.6.0",
     "rollup": "^1.31.0",
+    "rollup-plugin-web-worker-loader": "^1.3.0",
     "serve": "^11.3.0",
     "typescript": "latest"
   },
   "dependencies": {
+    "@typescript/vfs": "^1.0.0",
     "tslib": "^1.10.0"
   }
 }
