+++ rollup.config.js
@@ -1,26 +1,38 @@
-import typescript from '@rollup/plugin-typescript'
-import node from '@rollup/plugin-node-resolve'
-import commonjs from '@rollup/plugin-commonjs'
-import json from '@rollup/plugin-json'
+import typescript from "@rollup/plugin-typescript";
+import node from "@rollup/plugin-node-resolve";
+import commonjs from "@rollup/plugin-commonjs";
+import json from "@rollup/plugin-json";
+import webWorkerLoader from "rollup-plugin-web-worker-loader";
 
 // You can have more root bundles by extending this array
-const rootFiles = ['index.ts']
+const rootFiles = ["index.ts"];
 
-export default rootFiles.map(name => {
+export default rootFiles.map((name) => {
   /** @type { import("rollup").RollupOptions } */
   const options = {
     input: `src/${name}`,
-    external: ['typescript'],
+    external: ["typescript", "fs", "path"],
     output: {
       name,
-      dir: 'dist',
-      format: 'amd',
+      dir: "dist",
+      format: "amd",
+      paths: {
+        typescript: "typescript-sandbox/index",
+        fs: "typescript-sandbox/index",
+        path: "typescript-sandbox/index",
+      },
     },
-    plugins: [typescript({ tsconfig: 'tsconfig.json' }), commonjs(), node(), json()],
-  }
+    plugins: [
+      // node({ browser: true }),
+      typescript({ tsconfig: "tsconfig.json" }),
+      commonjs(),
+      json(),
+      webWorkerLoader(),
+    ],
+  };
 
-  return options
-})
+  return options;
+});
 
 /** Note:
  * if you end up wanting to import a dependency which relies on typescript, you will need
